requirements.txt
pyyaml==3.11
blessings==1.6
docopt==0.6.2
mkdocs==0.11.1
requests==2.5.0
# Tests
nose==1.3.4
responses==0.3.0
